Implemented normal mob spawning

+++ src/main/java/org/clawd/data/mobs/BossMob.java
@@ -16,7 +16,7 @@ public class BossMob extends NormalMob {
             String imgPath,
             double spawnChance,
             double xpDrop,
-            double goldDrop,
+            int goldDrop,
             boolean specialDrop,
             double health
     ) {
